Top-right Refresh button refreshing Manage, Invest and Marketplace pages
