+++ src/app/investor/dashboard/dashboard.component.html
@@ -53,7 +53,7 @@
         <div class="col-xs-1 col-sm-1 col-md-1">Actions</div>
       </div>
 
-        <app-dashboard-investment [assets]="assets"></app-dashboard-investment>
+        <app-dashboard-investment [assets]="this.dashboardService.getAssets()"></app-dashboard-investment>
 
     </div><!-- END MAIN Content  -->
