+++ src/app/investor/investor.component.html
@@ -33,7 +33,7 @@
           <p class="sw-fiat-value sw-index-value"> Index {{ETHprice}}</p>
         </div>
         <div class="col-xs-2 col-sm-2 col-md-2 left-border">
-          <a class="sw-button sw-btn-ghost1" (click)="refresh()">Refresh</a>
+          <a class="sw-button sw-btn-ghost1" (click)="refresh()" role="button">Refresh</a>
         </div>
         <div class="col-xs-1 col-sm-1 col-md-1">
           <img src="assets/img/icon-32.png" alt="MetaMask logo" class="icon" />
